+++ test/plug/accesslog/formatter_test.exs
@@ -65,6 +65,17 @@ defmodule Plug.AccessLog.FormatterTest do
   end
 
   test "%u" do
-    assert "-" == Formatter.format("%u", nil)
+    user        = "ex_unit"
+    credentials = Base.encode64("#{ user }:some_password")
+
+    empty      = conn(:get, "/")
+    valid      = empty |> put_req_header("Authorization", "Basic #{ credentials }")
+    incomplete = empty |> put_req_header("Authorization", "Basic #{ Base.encode64(user) }")
+    garbage    = empty |> put_req_header("Authorization", "Basic garbage")
+
+    assert "-"  == Formatter.format("%u", empty)
+    assert user == Formatter.format("%u", valid)
+    assert "-"  == Formatter.format("%u", incomplete)
+    assert "-"  == Formatter.format("%u", garbage)
   end
 end